web_scraping/requirements.txt
playwright==1.47.0
beautifulsoup4==4.12.3
readability-lxml==0.8.1
lxml==5.3.0
pydantic==2.9.2
urllib3==2.2.3
tldextract==5.1.2
python-slugify==8.0.4
requests==2.32.3
orjson==3.10.7
diskcache==5.6.3
sentence-transformers==3.0.1      # quick text embeddings for demo
faiss-cpu==1.8.0.post1
streamlit==1.38.0
